+++ http4k-contract/src/main/kotlin/org/http4k/contract/Security.kt
@@ -1,6 +1,7 @@
 package org.http4k.contract
 
 import org.http4k.core.Filter
+import org.http4k.core.Method
 import org.http4k.core.Request
 import org.http4k.core.Response
 import org.http4k.core.Status.Companion.UNAUTHORIZED
@@ -24,17 +25,22 @@ object NoSecurity : Security() {
 
 /**
  * Checks the presence of the named Api Key parameter. Filter returns 401 if Api-Key is not found in request.
+ * Includes an option to NOT authorise OPTIONS requests, which is currently not enabled for Swagger.
  */
-data class ApiKey<out T>(val param: Lens<Request, T>, private val validateKey: (T) -> Boolean) : Security() {
+data class ApiKey<out T>(val param: Lens<Request, T>, private val validateKey: (T) -> Boolean, private val authorizeOptionsRequests: Boolean = true) : Security() {
     override val filter = Filter {
         next ->
         {
-            val keyValid = try {
-                validateKey(param(it))
-            } catch (e: LensFailure) {
-                false
+            if(!authorizeOptionsRequests && it.method == Method.OPTIONS) {
+                next(it)
+            } else {
+                val keyValid = try {
+                    validateKey(param(it))
+                } catch (e: LensFailure) {
+                    false
+                }
+                if (keyValid) next(it) else Response(UNAUTHORIZED)
             }
-            if (keyValid) next(it) else Response(UNAUTHORIZED)
         }
     }
 }