+++ http4k-client-okhttp/src/test/kotlin/org/http4k/client/OkHttpStreamingTest.kt
@@ -3,5 +3,5 @@ package org.http4k.client
 import org.http4k.core.BodyMode
 import org.http4k.server.SunHttp
 
-class OkHttpStreamingTest : Http4kClientContract({ SunHttp(it) }, OkHttp(bodyMode = BodyMode.Stream),
+class OkHttpStreamingTest : HttpClientContract({ SunHttp(it) }, OkHttp(bodyMode = BodyMode.Stream),
     OkHttp(timeout, bodyMode = BodyMode.Stream))