+++ http4k-client-okhttp/src/test/kotlin/org/http4k/client/OkHttpTest.kt
@@ -2,4 +2,4 @@ package org.http4k.client
 
 import org.http4k.server.SunHttp
 
-class OkHttpTest : Http4kClientContract({ SunHttp(it) }, OkHttp(), OkHttp(timeout))
+class OkHttpTest : HttpClientContract({ SunHttp(it) }, OkHttp(), OkHttp(timeout))