+++ http4k-core/src/test/kotlin/org/http4k/core/http4kTest.kt
@@ -2,19 +2,20 @@ package org.http4k.core
 
 import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.equalTo
+import org.http4k.client.withAsyncApi
 import org.http4k.core.Status.Companion.OK
 import org.junit.Test
 import java.util.concurrent.CountDownLatch
 
-class AsyncTest {
+class AsyncHttpClientTest {
     @Test
-    fun `can convert a synchronous HttpHandler to mimic the async API`() {
+    fun `can convert a synchronous HttpHandler to mimic the AsyncHttpClient API`() {
 
         val handler = { _: Request -> Response(OK) }.withAsyncApi()
 
         val latch = CountDownLatch(1)
-        handler(Request(Method.GET, "/")) { response ->
-            assertThat(response.status, equalTo(OK))
+        handler(Request(Method.GET, "/")) {
+            assertThat(it.status, equalTo(OK))
             latch.countDown()
         }